Conflicts:
	src/backend/bridge/ddl.cpp
	src/backend/storage/data_table.cpp

Former-commit-id: f6f0d3ead6ccd1a1a8506ba556d02dc09e4463b1

+++ tools/pg_psql/help.cpp
@@ -477,12 +477,12 @@ helpSQL(const char *topic, unsigned short int pager)
 								 * word(s) */
 			{
 				wordlen = j = 1;
-				while (topic[j] != ' ' && j++ < len)
+				while (topic[j] != ' ' && j++ < static_cast<int>(len))
 					wordlen++;
 				if (x == 2)
 				{
 					j++;
-					while (topic[j] != ' ' && j++ <= len)
+					while (topic[j] != ' ' && j++ <= static_cast<int>(len))
 						wordlen++;
 				}
 				if (wordlen >= len)		/* Don't try again if the same word */
